add POC for traverse and index
